优化：课程id转为雪花id

+++ Sourcecode/Song.Entities/Guide.cs
@@ -12,7 +12,7 @@ namespace Song.Entities {
     		
     		protected String _Gc_Title;
     		
-    		protected Int32 _Cou_ID;
+    		protected Int64 _Cou_ID;
     		
     		protected String _Cou_Name;
     		
@@ -120,7 +120,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Cou_ID {
+    		public Int64 Cou_ID {
     			get {
     				return this._Cou_ID;
     			}
@@ -699,7 +699,7 @@ namespace Song.Entities {
     				this._Gc_Title = reader.GetString(_.Gc_Title);
     			}
     			if ((false == reader.IsDBNull(_.Cou_ID))) {
-    				this._Cou_ID = reader.GetInt32(_.Cou_ID);
+    				this._Cou_ID = reader.GetInt64(_.Cou_ID);
     			}
     			if ((false == reader.IsDBNull(_.Cou_Name))) {
     				this._Cou_Name = reader.GetString(_.Cou_Name);
@@ -867,7 +867,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Gc_Title = new WeiSha.Data.Field<Guide>("Gc_Title");
     			
     			/// <summary>
-    			/// 字段名：Cou_ID - 数据类型：Int32
+    			/// 字段名：Cou_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Cou_ID = new WeiSha.Data.Field<Guide>("Cou_ID");
