+++ Sourcecode/Song.Entities/Knowledge.cs
@@ -64,7 +64,7 @@ namespace Song.Entities {
     		
     		protected String _Org_Name;
     		
-    		protected Int32 _Cou_ID;
+    		protected Int64 _Cou_ID;
     		
     		protected String _Kns_UID;
     		
@@ -426,7 +426,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Cou_ID {
+    		public Int64 Cou_ID {
     			get {
     				return this._Cou_ID;
     			}
@@ -631,7 +631,7 @@ namespace Song.Entities {
     				this._Org_Name = reader.GetString(_.Org_Name);
     			}
     			if ((false == reader.IsDBNull(_.Cou_ID))) {
-    				this._Cou_ID = reader.GetInt32(_.Cou_ID);
+    				this._Cou_ID = reader.GetInt64(_.Cou_ID);
     			}
     			if ((false == reader.IsDBNull(_.Kns_UID))) {
     				this._Kns_UID = reader.GetString(_.Kns_UID);
@@ -803,7 +803,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Org_Name = new WeiSha.Data.Field<Knowledge>("Org_Name");
     			
     			/// <summary>
-    			/// 字段名：Cou_ID - 数据类型：Int32
+    			/// 字段名：Cou_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Cou_ID = new WeiSha.Data.Field<Knowledge>("Cou_ID");
