+++ Sourcecode/Song.Entities/Student_Ques.cs
@@ -12,7 +12,7 @@ namespace Song.Entities {
     		
     		protected Int32 _Ac_ID;
     		
-    		protected Int32 _Cou_ID;
+    		protected Int64 _Cou_ID;
     		
     		protected Int64 _Qus_ID;
     		
@@ -46,7 +46,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Cou_ID {
+    		public Int64 Cou_ID {
     			get {
     				return this._Cou_ID;
     			}
@@ -181,7 +181,7 @@ namespace Song.Entities {
     				this._Ac_ID = reader.GetInt32(_.Ac_ID);
     			}
     			if ((false == reader.IsDBNull(_.Cou_ID))) {
-    				this._Cou_ID = reader.GetInt32(_.Cou_ID);
+    				this._Cou_ID = reader.GetInt64(_.Cou_ID);
     			}
     			if ((false == reader.IsDBNull(_.Qus_ID))) {
     				this._Qus_ID = reader.GetInt64(_.Qus_ID);
@@ -238,7 +238,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Ac_ID = new WeiSha.Data.Field<Student_Ques>("Ac_ID");
     			
     			/// <summary>
-    			/// 字段名：Cou_ID - 数据类型：Int32
+    			/// 字段名：Cou_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Cou_ID = new WeiSha.Data.Field<Student_Ques>("Cou_ID");
