+++ Sourcecode/Song.Entities/TestResults.cs
@@ -56,7 +56,7 @@ namespace Song.Entities {
     		
     		protected String _St_IDCardNumber;
     		
-    		protected Int32 _Cou_ID;
+    		protected Int64 _Cou_ID;
     		
     		/// <summary>
     		/// -1
@@ -340,7 +340,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Cou_ID {
+    		public Int64 Cou_ID {
     			get {
     				return this._Cou_ID;
     			}
@@ -513,7 +513,7 @@ namespace Song.Entities {
     				this._St_IDCardNumber = reader.GetString(_.St_IDCardNumber);
     			}
     			if ((false == reader.IsDBNull(_.Cou_ID))) {
-    				this._Cou_ID = reader.GetInt32(_.Cou_ID);
+    				this._Cou_ID = reader.GetInt64(_.Cou_ID);
     			}
     		}
     		
@@ -662,7 +662,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field St_IDCardNumber = new WeiSha.Data.Field<TestResults>("St_IDCardNumber");
     			
     			/// <summary>
-    			/// 字段名：Cou_ID - 数据类型：Int32
+    			/// 字段名：Cou_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Cou_ID = new WeiSha.Data.Field<TestResults>("Cou_ID");
     		}
