+++ Sourcecode/Song.ServiceImpls/OutlineCom.cs
@@ -98,7 +98,7 @@ namespace Song.ServiceImpls
         /// <param name="couid">课程（或课程）id</param>
         /// <param name="names">名称，可以是用逗号分隔的多个名称</param>
         /// <returns></returns>
-        public Outline OutlineBatchAdd(int orgid, int sbjid, int couid, string names)
+        public Outline OutlineBatchAdd(int orgid, int sbjid, long couid, string names)
         {
             //整理名称信息
             names = names.Replace("，", ",");
@@ -140,7 +140,7 @@ namespace Song.ServiceImpls
         /// <param name="pid">上级id</param>
         /// <param name="name"></param>
         /// <returns></returns>
-        public Outline OutlineIsExist(int orgid, int sbjid, int couid, long pid, string name)
+        public Outline OutlineIsExist(int orgid, int sbjid, long couid, long pid, string name)
         {
             WhereClip wc = Outline._.Org_ID == orgid;
             if (sbjid > 0) wc &= Outline._.Sbj_ID == sbjid;
@@ -243,7 +243,7 @@ namespace Song.ServiceImpls
         /// <param name="path"></param>
         /// <param name="couid">课程ID</param>
         /// <returns></returns>
-        public string OutlineExport4Excel(string path, int couid)
+        public string OutlineExport4Excel(string path, long couid)
         {
             HSSFWorkbook hssfworkbook = new HSSFWorkbook();
             //xml配置文件
@@ -419,7 +419,7 @@ namespace Song.ServiceImpls
         /// <param name="couid">课程ID</param>
         /// <param name="names">多级名称</param>
         /// <returns></returns>
-        public Outline OutlineSingle(int couid, List<string> names)
+        public Outline OutlineSingle(long couid, List<string> names)
         {
             Outline curr = null;
             int index = 0;
@@ -466,7 +466,7 @@ namespace Song.ServiceImpls
         /// <param name="couid">课程Id</param>
         /// <param name="isUse">是否包括只是允许的章节,null取所有范围，true只是允许采用的章节,false反之</param>
         /// <returns></returns>
-        public Outline OutlineFirst(int couid, bool? isUse)
+        public Outline OutlineFirst(long couid, bool? isUse)
         {
             Song.Entities.Outline[] outlines = null;
             WhereClip wc = Outline._.Cou_ID == couid;
@@ -515,7 +515,7 @@ namespace Song.ServiceImpls
         /// <param name="couid">所属课程id</param>
         /// <param name="isUse"></param>
         /// <returns></returns>
-        public Outline[] OutlineAll(int couid, bool? isUse)
+        public Outline[] OutlineAll(long couid, bool? isUse)
         {
             ////从缓存中读取
             //try
@@ -613,7 +613,7 @@ namespace Song.ServiceImpls
         /// </summary>
         /// <param name="couid">课程ID</param>
         /// <returns></returns>
-        public int OutlineCleanup(int couid)
+        public int OutlineCleanup(long couid)
         {
             WhereClip wc = new WhereClip();
             if (couid > 0) wc &= Outline._.Cou_ID == couid;
@@ -684,7 +684,7 @@ namespace Song.ServiceImpls
         /// <param name="isUse"></param>
         /// <param name="count">取多少条记录，如果小于等于0，则取所有</param>
         /// <returns></returns>
-        public Outline[] OutlineCount(int couid, string search, bool? isUse, int count)
+        public Outline[] OutlineCount(long couid, string search, bool? isUse, int count)
         {
             return OutlineCount(couid, null, search, isUse, count);
         }
@@ -728,7 +728,7 @@ namespace Song.ServiceImpls
         /// <param name="isUse"></param>
         /// <param name="count"></param>
         /// <returns></returns>
-        public Outline[] OutlineCount(int couid, bool? islive, string search, bool? isUse, int count)
+        public Outline[] OutlineCount(long couid, bool? islive, string search, bool? isUse, int count)
         {
             WhereClip wc = Outline._.Cou_ID == couid;
             if (!string.IsNullOrWhiteSpace(search)) wc.And(Outline._.Ol_Name.Like("%" + search + "%"));
@@ -736,7 +736,7 @@ namespace Song.ServiceImpls
             if (isUse != null) wc.And(Outline._.Ol_IsUse == (bool)isUse);
             return Gateway.Default.From<Outline>().Where(wc).OrderBy(Outline._.Ol_Tax.Asc).ToArray<Outline>(count);
         }
-        public Outline[] OutlineCount(int couid, long pid, bool? isUse, int count)
+        public Outline[] OutlineCount(long couid, long pid, bool? isUse, int count)
         {
             //List<Outline> tm = null;
             //try
@@ -769,7 +769,7 @@ namespace Song.ServiceImpls
             if (isUse != null) wc.And(Outline._.Ol_IsUse == (bool)isUse);
             return Gateway.Default.From<Outline>().Where(wc).OrderBy(Outline._.Ol_Tax.Asc).ToArray<Outline>(count);
         }
-        public Outline[] OutlineCount(int orgid, int sbjid, int couid, long pid, bool? isUse, int count)
+        public Outline[] OutlineCount(int orgid, int sbjid, long couid, long pid, bool? isUse, int count)
         {
             WhereClip wc = new WhereClip();
             if (orgid > 0) wc &= Outline._.Org_ID == orgid;
@@ -786,14 +786,14 @@ namespace Song.ServiceImpls
         /// <param name="pid"></param>
         /// <param name="isUse"></param>
         /// <returns></returns>
-        public int OutlineOfCount(int couid, long pid, bool? isUse)
+        public int OutlineOfCount(long couid, long pid, bool? isUse)
         {
             WhereClip wc = Outline._.Cou_ID == couid;
             if (pid >= 0) wc.And(Outline._.Ol_PID == pid);
             if (isUse != null) wc.And(Outline._.Ol_IsUse == (bool)isUse);
             return Gateway.Default.Count<Outline>(wc);
         }
-        public int OutlineOfCount(int couid, long pid, bool? isUse, bool children)
+        public int OutlineOfCount(long couid, long pid, bool? isUse, bool children)
         {
             if (!children) return this.OutlineOfCount(couid, pid, isUse);
 
@@ -816,7 +816,7 @@ namespace Song.ServiceImpls
         /// <param name="isUse">是否启用</param>
         /// <param name="isVideo">是否有视频</param>
         /// <returns></returns>
-        public int OutlineOfCount(int couid, long pid, bool? isUse, bool? isVideo, bool? isFinish)
+        public int OutlineOfCount(long couid, long pid, bool? isUse, bool? isVideo, bool? isFinish)
         {
             WhereClip wc = new WhereClip();
             if (couid > 0) wc &= Outline._.Cou_ID == couid;
@@ -833,7 +833,7 @@ namespace Song.ServiceImpls
         /// <param name="pid">父id</param>
         /// <param name="isUse"></param>
         /// <returns></returns>
-        public bool OutlineIsChildren(int couid, long pid, bool? isUse)
+        public bool OutlineIsChildren(long couid, long pid, bool? isUse)
         {
             WhereClip wc = Outline._.Cou_ID == couid;
             if (pid >= 0) wc.And(Outline._.Ol_PID == pid);
@@ -861,7 +861,7 @@ namespace Song.ServiceImpls
         /// <param name="pid"></param>
         /// <param name="isUse"></param>
         /// <returns></returns>
-        public Outline[] OutlineChildren(int couid, long pid, bool? isUse,int count)
+        public Outline[] OutlineChildren(long couid, long pid, bool? isUse,int count)
         {
             WhereClip wc = Outline._.Cou_ID == couid;
             if (pid >= 0) wc.And(Outline._.Ol_PID == pid);
@@ -878,7 +878,7 @@ namespace Song.ServiceImpls
         /// <param name="index"></param>
         /// <param name="countSum"></param>
         /// <returns></returns>
-        public Outline[] OutlinePager(int couid, bool? isUse, string searTxt, int size, int index, out int countSum)
+        public Outline[] OutlinePager(long couid, bool? isUse, string searTxt, int size, int index, out int countSum)
         {
             WhereClip wc = Outline._.Cou_ID == couid;
             if (isUse != null) wc.And(Outline._.Ol_IsUse == (bool)isUse);
@@ -1071,7 +1071,7 @@ namespace Song.ServiceImpls
         /// <param name="type">事件类型，1为提醒，2为知识展示，3课堂提问，4实时反馈（例如，选择某项后跳转到某秒）</param>
         /// <param name="isUse"></param>
         /// <returns></returns>
-        public OutlineEvent[] EventAll(int couid, long olid, int type, bool? isUse)
+        public OutlineEvent[] EventAll(long couid, long olid, int type, bool? isUse)
         {
             WhereClip wc = OutlineEvent._.Ol_ID == olid;
             if (couid > 0) wc.And(OutlineEvent._.Oe_IsUse == (bool)isUse);
@@ -1087,7 +1087,7 @@ namespace Song.ServiceImpls
         /// <param name="type"></param>
         /// <param name="isUse"></param>
         /// <returns></returns>
-        public OutlineEvent[] EventAll(int couid, string uid, int type, bool? isUse)
+        public OutlineEvent[] EventAll(long couid, string uid, int type, bool? isUse)
         {
             WhereClip wc = OutlineEvent._.Ol_UID == uid;
             if (couid > 0) wc.And(OutlineEvent._.Oe_IsUse == (bool)isUse);
